change key_size size

+++ kpayload/include/ccp_helper.h
@@ -27,11 +27,11 @@ union ccp_op {
 		uint64_t data;//0x10
 		uint64_t data_size_bits;//0x18
 		uint8_t hash[0x20]; //0x20
-		union {
+		union{
 			uint32_t key_index;
 			uint8_t key[0x40];
 		};
-		uint64_t key_size;
+		uint32_t key_size;
 
 	} hmac;
